move a few things to pep, import from pep, tidy a test module, rename looper-specific project type module
